requirements.txt
streamlit
openai
azure-ai-formrecognizer
